refactor: extract utilities
