+++ src/components/BoxweaverCalculator.js
@@ -1,205 +1,7 @@
-import React, { useState, useCallback } from 'react';
+import React, { useState } from 'react';
 import { Upload, Download, Play, Box, FileSpreadsheet, Settings } from 'lucide-react';
 import * as XLSX from 'xlsx';
-
-// 3D Bin Packing Implementation
-class Item3D {
-  constructor(name, width, height, depth, weight) {
-    this.name = name;
-    this.width = width;
-    this.height = height;
-    this.depth = depth;
-    this.weight = weight;
-    this.position = null;
-    this.rotationType = 0;
-  }
-
-  getDimension(axis) {
-    const dimensions = [this.width, this.height, this.depth];
-    const rotations = [
-      [0, 1, 2], // No rotation
-      [0, 2, 1], // Rotate around X
-      [1, 0, 2], // Rotate around Y
-      [1, 2, 0], // Rotate around X then Y
-      [2, 0, 1], // Rotate around Z
-      [2, 1, 0]  // Rotate around Z then Y
-    ];
-
-    const rotation = rotations[this.rotationType];
-    return dimensions[rotation[axis]];
-  }
-
-  getWidth() { return this.getDimension(0); }
-  getHeight() { return this.getDimension(1); }
-  getDepth() { return this.getDimension(2); }
-}
-
-class Position {
-  constructor(x, y, z) {
-    this.x = x;
-    this.y = y;
-    this.z = z;
-  }
-}
-
-class Bin3D {
-  constructor(name, width, height, depth, maxWeight) {
-    this.name = name;
-    this.width = width;
-    this.height = height;
-    this.depth = depth;
-    this.maxWeight = maxWeight;
-    this.items = [];
-    this.currentWeight = 0;
-  }
-
-  reset() {
-    this.items = [];
-    this.currentWeight = 0;
-  }
-
-  canAddItem(item) {
-    return this.currentWeight + item.weight <= this.maxWeight;
-  }
-
-  findBestPosition(item) {
-    if (!this.canAddItem(item)) return null;
-
-    // Try all rotation types
-    for (let rotation = 0; rotation < 6; rotation++) {
-      item.rotationType = rotation;
-
-      // Get item dimensions after rotation
-      const itemWidth = item.getWidth();
-      const itemHeight = item.getHeight();
-      const itemDepth = item.getDepth();
-
-      // Check if item fits in bin dimensions
-      if (itemWidth > this.width || itemHeight > this.height || itemDepth > this.depth) {
-        continue;
-      }
-
-      // Find all possible positions
-      const positions = this.getPossiblePositions(itemWidth, itemHeight, itemDepth);
-
-      // Find the best position (lowest and most to the back-left)
-      let bestPosition = null;
-      let minY = Infinity;
-
-      for (const pos of positions) {
-        if (this.canPlaceItem(pos, itemWidth, itemHeight, itemDepth)) {
-          if (pos.y < minY || (pos.y === minY && (pos.z < bestPosition.z || (pos.z === bestPosition.z && pos.x < bestPosition.x)))) {
-            bestPosition = pos;
-            minY = pos.y;
-          }
-        }
-      }
-
-      if (bestPosition) {
-        item.position = bestPosition;
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  getPossiblePositions(itemWidth, itemHeight, itemDepth) {
-    const positions = [new Position(0, 0, 0)]; // Always try origin
-
-    // Add positions based on existing items
-    for (const placedItem of this.items) {
-      const x = placedItem.position.x;
-      const y = placedItem.position.y;
-      const z = placedItem.position.z;
-      const w = placedItem.getWidth();
-      const h = placedItem.getHeight();
-      const d = placedItem.getDepth();
-
-      // Try positions adjacent to each placed item
-      positions.push(new Position(x + w, y, z)); // Right
-      positions.push(new Position(x, y + h, z)); // Top
-      positions.push(new Position(x, y, z + d)); // Front
-    }
-
-    // Filter positions that are within bin bounds
-    return positions.filter(pos => 
-      pos.x + itemWidth <= this.width &&
-      pos.y + itemHeight <= this.height &&
-      pos.z + itemDepth <= this.depth
-    );
-  }
-
-  canPlaceItem(position, width, height, depth) {
-    // Check if the item overlaps with any existing items
-    for (const item of this.items) {
-      if (this.intersects(
-        position.x, position.y, position.z, width, height, depth,
-        item.position.x, item.position.y, item.position.z,
-        item.getWidth(), item.getHeight(), item.getDepth()
-      )) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  intersects(x1, y1, z1, w1, h1, d1, x2, y2, z2, w2, h2, d2) {
-    return !(x1 + w1 <= x2 || x2 + w2 <= x1 ||
-             y1 + h1 <= y2 || y2 + h2 <= y1 ||
-             z1 + d1 <= z2 || z2 + d2 <= z1);
-  }
-
-  addItem(item) {
-    if (this.findBestPosition(item)) {
-      this.items.push(item);
-      this.currentWeight += item.weight;
-      return true;
-    }
-    return false;
-  }
-}
-
-// Binary search to find maximum items that fit
-function findMaxCapacity(bin, itemTemplate, maxAttempts = 1000) {
-  let low = 0;
-  let high = maxAttempts;
-  let bestFit = 0;
-
-  // First check weight constraint
-  const maxByWeight = Math.floor(bin.maxWeight / itemTemplate.weight);
-  high = Math.min(high, maxByWeight);
-
-  while (low <= high) {
-    const mid = Math.floor((low + high) / 2);
-    bin.reset();
-
-    let allFit = true;
-    for (let i = 0; i < mid; i++) {
-      const item = new Item3D(
-        `${itemTemplate.name}_${i}`,
-        itemTemplate.width,
-        itemTemplate.height,
-        itemTemplate.depth,
-        itemTemplate.weight
-      );
-
-      if (!bin.addItem(item)) {
-        allFit = false;
-        break;
-      }
-    }
-
-    if (allFit) {
-      bestFit = mid;
-      low = mid + 1;
-    } else {
-      high = mid - 1;
-    }
-  }
-
-  return bestFit;
-}
+import { processFile } from '../utils/fileProcessor';
 
 // Main component
 export default function BoxCapacityCalculator() {
@@ -252,78 +54,20 @@ export default function BoxCapacityCalculator() {
     }
   };
 
-  const processFile = async () => {
+  const handleProcessFile = async () => {
     if (!file) return;
 
     setIsProcessing(true);
     setProgress(0);
 
     try {
-      // Read the file
-      const data = await file.arrayBuffer();
-      const workbook = XLSX.read(data);
-
-      // Extract data from sheets
-      const itemsSheet = workbook.Sheets['Items'];
-      const boxesSheet = workbook.Sheets['Boxes'];
-
-      if (!itemsSheet || !boxesSheet) {
-        throw new Error('Missing required sheets: Items and/or Boxes');
-      }
-
-      const items = XLSX.utils.sheet_to_json(itemsSheet);
-      const boxes = XLSX.utils.sheet_to_json(boxesSheet);
-
-      const totalOperations = items.length * boxes.length;
-      let currentOperation = 0;
-
-      // Process each item
-      const processedResults = [];
-
-      for (const item of items) {
-        const result = {
-          SKU: item.SKU,
-          Height: item.Height,
-          Width: item.Width,
-          Length: item.Length,
-          Weight: item.Weight,
-        };
-
-        // Calculate capacity for each box type
-        for (const box of boxes) {
-          currentOperation++;
-          setProgress((currentOperation / totalOperations) * 100);
-
-          // Apply tolerances
-          const adjustedBox = new Bin3D(
-            box.BoxType,
-            box.Width - dimensionTolerance,
-            box.Height - dimensionTolerance,
-            box.Length - dimensionTolerance,
-            box.MaxWeight - weightTolerance
-          );
-
-          const itemTemplate = {
-            name: item.SKU,
-            width: item.Width,
-            height: item.Height,
-            depth: item.Length,
-            weight: item.Weight
-          };
-
-          // Find maximum capacity using binary search
-          const capacity = findMaxCapacity(adjustedBox, itemTemplate);
-          result[`Units_in_${box.BoxType}`] = capacity;
-
-          // Small delay to show progress
-          await new Promise(resolve => setTimeout(resolve, 10));
-        }
-
-        processedResults.push(result);
-      }
-
+      const processedResults = await processFile(
+        file,
+        dimensionTolerance,
+        weightTolerance,
+        setProgress
+      );
       setResults(processedResults);
-      setProgress(100);
     } catch (error) {
       alert(`Error processing file: ${error.message}`);
     } finally {
@@ -463,7 +207,7 @@ export default function BoxCapacityCalculator() {
           {/* Run Button */}
           {file && !isProcessing && (
             <button
-              onClick={processFile}
+              onClick={handleProcessFile}
               className="w-full flex items-center justify-center gap-2 text-white px-6 py-4 rounded-lg transition-colors text-lg font-semibold"
               style={{ backgroundColor: '#F5683A' }}
               onMouseEnter={(e) => e.target.style.backgroundColor = '#E55529'}
